+++ src/main/kotlin/com/coder/gateway/models/UriScheme.kt
@@ -1,0 +1,5 @@
+package com.coder.gateway.models
+
+enum class UriScheme(val scheme: String) {
+    HTTP("http"), HTTPS("https")
+}