+++ src/main/kotlin/com/jetbrains/gateway/sdk/convertors/InstantConverter.kt
@@ -1,0 +1,65 @@
+package com.jetbrains.gateway.sdk.convertors
+
+import com.google.gson.JsonDeserializationContext
+import com.google.gson.JsonDeserializer
+import com.google.gson.JsonElement
+import com.google.gson.JsonParseException
+import com.google.gson.JsonPrimitive
+import com.google.gson.JsonSerializationContext
+import com.google.gson.JsonSerializer
+import java.lang.reflect.Type
+import java.time.Instant
+import java.time.format.DateTimeFormatter
+import java.time.temporal.TemporalAccessor
+
+/**
+ * GSON serialiser/deserialiser for converting [Instant] objects.
+ */
+class InstantConverter : JsonSerializer<Instant?>, JsonDeserializer<Instant?> {
+    /**
+     * Gson invokes this call-back method during serialization when it encounters a field of the
+     * specified type.
+     *
+     *
+     *
+     * In the implementation of this call-back method, you should consider invoking
+     * [JsonSerializationContext.serialize] method to create JsonElements for any
+     * non-trivial field of the `src` object. However, you should never invoke it on the
+     * `src` object itself since that will cause an infinite loop (Gson will call your
+     * call-back method again).
+     *
+     * @param src the object that needs to be converted to Json.
+     * @param typeOfSrc the actual type (fully genericized version) of the source object.
+     * @return a JsonElement corresponding to the specified object.
+     */
+    override fun serialize(src: Instant?, typeOfSrc: Type?, context: JsonSerializationContext?): JsonElement? {
+        return JsonPrimitive(FORMATTER.format(src))
+    }
+
+    /**
+     * Gson invokes this call-back method during deserialization when it encounters a field of the
+     * specified type.
+     *
+     *
+     *
+     * In the implementation of this call-back method, you should consider invoking
+     * [JsonDeserializationContext.deserialize] method to create objects
+     * for any non-trivial field of the returned object. However, you should never invoke it on the
+     * the same type passing `json` since that will cause an infinite loop (Gson will call your
+     * call-back method again).
+     *
+     * @param json The Json data being deserialized
+     * @param typeOfT The type of the Object to deserialize to
+     * @return a deserialized object of the specified type typeOfT which is a subclass of `T`
+     * @throws JsonParseException if json is not in the expected format of `typeOfT`
+     */
+    @Throws(JsonParseException::class)
+    override fun deserialize(json: JsonElement, typeOfT: Type, context: JsonDeserializationContext): Instant {
+        return FORMATTER.parse(json.asString) { temporal: TemporalAccessor? -> Instant.from(temporal) }
+    }
+
+    companion object {
+        /** Formatter.  */
+        private val FORMATTER = DateTimeFormatter.ISO_INSTANT
+    }
+}